refactor: move Header to layout and add board/Column component
