+++ .gitignore
@@ -8,6 +8,6 @@ dist/
 .env
 .env.local
 .env.*.local
-
+.envfolders
 # misc
 .DS_Store
